refactor(cli): extract platform installers from __main__ to cli/installers
graphify/__main__.py had grown to 4582 lines, mostly because every
supported AI platform (Claude Code, Gemini, VSCode, Kiro, Antigravity,
Cursor, Devin, Kilo, OpenCode, Codex, Amp, CodeBuddy) ships its own
install/uninstall code path. This is the Phase 5.1 refactor the long
plan called for: move the installer block (lines 1-2069 + _clone_repo)
out into graphify/cli/installers.py.

Why this split is safe:

1. Backward compatibility preserved by re-exporting every installer
   name as a module attribute on __main__ (set via sys.modules), so
   existing imports like `from graphify.__main__ import claude_install`
   and `from graphify.__main__ import _CLAUDE_MD_SECTION` keep working.
2. The PEP 562 `_ALWAYS_ON_ALIASES` lazy block (which doesn't show up
   in dir()) is also bound to the __main__ module, so install-string
   tests pass without modification.
3. All Path(__file__).parent references in installers.py are bumped
   to .parent.parent so packaged assets (always_on/*.md, skills/*,
   command-kilo.md, skill-{vscode,copilot}.md) keep resolving from
   graphify/ rather than the new graphify/cli/ location.

Results:

- __main__.py: 4582 -> 2617 lines (-43%, 1965 LOC out)
- New graphify/cli/{__init__.py,installers.py}: 2090 lines
- pytest: 1968 passed, 1 skipped (no regressions)
- ruff: same 1 pre-existing F821 in prs.py, zero new errors
- graphify --version / --help / test_install_* / test_claude_md /
  test_install_roundtrip all pass
